% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{de_analysis}
\alias{de_analysis}
\title{Perform differential expression analysis}
\usage{
de_analysis(
  counts_file,
  coldata_file,
  output_dir,
  target,
  reference_level,
  contrast_level,
  covariates,
  lfc_thresh = 1,
  pval_thresh = 0.05,
  fdr = 0.05
)
}
\arguments{
\item{counts_file}{Path to the counts file}

\item{coldata_file}{Path to the coldata file}

\item{output_dir}{Directory to save results}

\item{target}{Target variable name}

\item{reference_level}{Reference level for the condition}

\item{contrast_level}{Contrast level for the condition}

\item{covariates}{List of covariates}

\item{lfc_thresh}{Log fold change threshold}

\item{pval_thresh}{P-value threshold}

\item{fdr}{False discovery rate}
}
\description{
This function performs differential expression analysis using DESeq2.
}
