% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_go.R
\name{gsea_go}
\alias{gsea_go}
\title{Perform GSEA GO analysis}
\usage{
gsea_go(de_file, gsea_directory)
}
\arguments{
\item{de_file}{Path to the DE results file}

\item{gsea_directory}{Directory to save GSEA results}
}
\description{
This function performs Gene Set Enrichment Analysis (GSEA) using Gene Ontology (GO) terms.
}
