% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst_transform.R
\name{vst_transform}
\alias{vst_transform}
\title{Perform variance stabilizing transformation (VST)}
\usage{
vst_transform(
  counts_file,
  coldata_file,
  output_dir,
  design_formula_cols = NULL
)
}
\arguments{
\item{counts_file}{Path to the counts file}

\item{coldata_file}{Path to the coldata file}

\item{output_dir}{Directory to save VST results}

\item{design_formula_cols}{Coldata fields to include in design formula}
}
\description{
This function performs VST on RNA-seq count data.
}
